Adapting baseapp_notifications to follow the plugin architecture
