+++ baseapp_blocks/services.py
@@ -7,7 +7,7 @@ from query_optimizer.typing import GQLInfo
 
 from baseapp_core.plugins import SharedServiceProvider
 
-# Hint key set by _exclude_blocked_profiles so get_queryset() knows
+# Hint key set by _exclude_blocked_from_foreign_queryset so get_queryset() knows
 # filtering was already applied and can skip a redundant .exclude().
 _BLOCKED_PROFILES_FILTERED_HINT = "_blocked_profiles_filtered"
 
@@ -20,7 +20,9 @@ class BlockLookupService(SharedServiceProvider):
     def is_available(self) -> bool:
         return apps.is_installed("baseapp_blocks")
 
-    def exclude_blocked_profiles(self, queryset: models.QuerySet, info: GQLInfo) -> models.QuerySet:
+    def exclude_blocked_from_foreign_queryset(
+        self, queryset: models.QuerySet, info: GQLInfo
+    ) -> models.QuerySet:
         """Exclude comments from blocked/blocking profiles.
 
         Must be called BEFORE evaluate_with_prefetch_hack / optimize
@@ -37,16 +39,23 @@ class BlockLookupService(SharedServiceProvider):
             queryset._hints[_BLOCKED_PROFILES_FILTERED_HINT] = True
             return queryset
 
-        profile = getattr(user, "current_profile", None)
-        if not profile:
-            queryset._hints[_BLOCKED_PROFILES_FILTERED_HINT] = True
-            return queryset
+        if apps.is_installed("baseapp_profiles"):
+            profile = getattr(user, "current_profile", None)
+            if not profile:
+                queryset._hints[_BLOCKED_PROFILES_FILTERED_HINT] = True
+                return queryset
 
-        blocked_profile_ids = profile.blocking.values_list("target_id", flat=True)
-        blocker_profile_ids = profile.blockers.values_list("actor_id", flat=True)
+            blocked_profile_ids = profile.blocking.values_list("target_id", flat=True)
+            blocker_profile_ids = profile.blockers.values_list("actor_id", flat=True)
 
-        qs = queryset.exclude(
-            Q(profile__id__in=blocked_profile_ids) | Q(profile__id__in=blocker_profile_ids)
-        )
+            qs = queryset.exclude(
+                Q(profile__id__in=blocked_profile_ids) | Q(profile__id__in=blocker_profile_ids)
+            )
+            qs._hints[_BLOCKED_PROFILES_FILTERED_HINT] = True
+            return qs
+
+        blocked_user_ids = user.social_blocks.values_list("target_id", flat=True)
+        blocker_user_ids = user.blockers.values_list("user_id", flat=True)
+        qs = qs.exclude(Q(user__id__in=blocked_user_ids) | Q(user__id__in=blocker_user_ids))
         qs._hints[_BLOCKED_PROFILES_FILTERED_HINT] = True
         return qs
