+++ baseapp_notifications/graphql/interfaces.py
@@ -1,0 +1,4 @@
+def get_notifications_interface():
+    from .object_types import NotificationsInterface
+
+    return NotificationsInterface
