+++ baseapp_notifications/tests/conftest.py
@@ -1,2 +1,4 @@
 from baseapp_core.graphql.testing.fixtures import *  # noqa
 from baseapp_core.tests.fixtures import *  # noqa
+
+from .fixtures import *  # noqa
