+++ baseapp_notifications/utils.py
@@ -6,87 +6,6 @@ from django.db.models import Q
 from django.template.exceptions import TemplateDoesNotExist
 from django.template.loader import render_to_string
 from django.utils.text import slugify
-from notifications.signals import notify
-
-from .tasks import send_push_notification
-
-
-def send_notification(
-    sender,
-    recipient,
-    verb,
-    description=None,
-    action_object=None,
-    target=None,
-    add_to_history=True,
-    send_push=True,
-    send_email=True,
-    email_subject=None,
-    email_message=None,
-    push_title=None,
-    push_description=None,
-    **kwargs,
-):
-    NotificationSetting = swapper.load_model("baseapp_notifications", "NotificationSetting")
-    notifications = []
-
-    if add_to_history and can_user_receive_notification(
-        recipient.id, verb, NotificationSetting.NotificationChannelTypes.IN_APP
-    ):
-        notifications = notify.send(
-            sender=sender,
-            recipient=recipient,
-            verb=verb,
-            action_object=action_object,
-            description=description,
-            target=target,
-            **kwargs,
-        )
-
-    if send_email and can_user_receive_notification(
-        recipient.id, verb, NotificationSetting.NotificationChannelTypes.EMAIL
-    ):
-        notification = (
-            notifications[0][1][0]
-            if len(notifications) > 0 and len(notifications[0]) > 1 and len(notifications[0][1]) > 0
-            else None
-        )
-
-        send_email_notification(
-            to=recipient.email,
-            context=dict(
-                notification=notification,
-                sender=sender,
-                recipient=recipient,
-                verb=verb,
-                action_object=action_object,
-                description=description,
-                target=target,
-                add_to_history=add_to_history,
-                send_push=send_push,
-                email_subject=email_subject or description,
-                email_message=email_message or description,
-                **kwargs,
-            ),
-        )
-
-        if notification:
-            notification.emailed = True
-            notification.save(update_fields=["emailed"])
-
-    if send_push and can_user_receive_notification(
-        recipient.id, verb, NotificationSetting.NotificationChannelTypes.PUSH
-    ):
-        send_push_notification.delay(
-            recipient.id,
-            push_title=push_title,
-            push_description=push_description or description,
-            # TO DO:
-            # serialize all objects so devices can use this data if necessary
-            **kwargs,
-        )
-
-    return notifications
 
 
 def render_verb_template_or_default(verb, context, template_type, extension):
