+++ baseapp_reactions/notifications.py
@@ -5,7 +5,7 @@ from celery import shared_task
 from django.contrib.auth import get_user_model
 from django.utils.translation import gettext as _
 
-from baseapp_notifications import send_notification
+from baseapp_core.plugins import shared_services
 
 Reaction = swapper.load_model("baseapp_reactions", "Reaction")
 User = get_user_model()
@@ -13,6 +13,10 @@ User = get_user_model()
 
 @shared_task
 def send_reaction_created_notification(reaction_pk, recipient_id):
+    service = shared_services.get("notifications")
+    if not service:
+        return
+
     try:
         reaction = Reaction.objects.get(pk=reaction_pk)
         recipient = User.objects.get(pk=recipient_id)
@@ -22,7 +26,7 @@ def send_reaction_created_notification(reaction_pk, recipient_id):
 
     sender = getattr(reaction, "profile", None) or reaction.user
 
-    send_notification(
+    service.send_notification(
         add_to_history=True,
         send_push=True,
         send_email=True,
