+++ pyproject.toml
@@ -169,13 +169,14 @@ baseapp_api_key = "baseapp_api_key.plugin:ApiKeyPlugin"
 baseapp_profiles = "baseapp_profiles.plugin:ProfilesPlugin"
 baseapp_pages = "baseapp_pages.plugin:PagesPlugin"
 baseapp_organizations = "baseapp_organizations.plugin:OrganizationsPlugin"
-baseapp_comments = "baseapp_comments.plugin:CommentsPlugin"
 baseapp_payments = "baseapp_payments.plugin:PaymentsPlugin"
 baseapp_referrals = "baseapp_referrals.plugin:ReferralsPlugin"
 baseapp_pdf = "baseapp_pdf.plugin:PDFPlugin"
 baseapp_message_templates = "baseapp_message_templates.plugin:MessageTemplatesPlugin"
 baseapp_cloudflare_stream_field = "baseapp_cloudflare_stream_field.plugin:CloudflareStreamFieldPlugin"
 baseapp_e2e = "baseapp_e2e.plugin:E2EPlugin"
+baseapp_notifications = "baseapp_notifications.plugin:NotificationsPlugin"
+baseapp_comments = "baseapp_comments.plugin:CommentsPlugin"
 
 [tool.black]
 line-length = 100
